+++ LogicBuilder.Kendo.ExpressionExtensions.IntegrationTests/Helpers.cs
@@ -173,9 +173,9 @@ namespace LogicBuilder.Kendo.ExpressionExtensions.IntegrationTests
 
             async Task<IEnumerable> GetData()
             {
-                var groupByExpressions = request.CreateGroupedByQueryExpressions<TModel>();
-                IQueryable<TModel> pagedQuery = await contextRepository.QueryAsync<TModel, TData, IQueryable<TModel>, IQueryable<TData>>(groupByExpressions.PagingExpression, selectExpandDefinition);
-                return groupByExpressions.GroupByExpression.Compile()(pagedQuery);
+                var groupByExpressions = request.CreateGroupingQueryExpressions<TModel>();
+                IQueryable<TModel> pagedQuery = await contextRepository.QueryAsync<TModel, TData, IQueryable<TModel>, IQueryable<TData>>(groupByExpressions.UngroupedUnpagedExpression, selectExpandDefinition);
+                return groupByExpressions.PagedGroupedExpression.Compile()(pagedQuery);
             }
 
             async Task<AggregateFunctionsGroup> GetAggregateFunctionsGroup()